+++ frontend/app/report_lost/page.tsx
@@ -58,7 +58,7 @@ const ReportLost = () => {
         const itemName = form.querySelector('textarea[name="itemName"]').value; // Retrieves the name from the textarea
         const businessId = form.querySelector('select').value; // Business ID
         const itemPhotoUrl = form.querySelector('input[type="url"]').value;
-        const description = form.querySelector('textarea').value;
+        const description = form.querySelector('textarea[name="description"]').value;
         const bnfUserId = session?.id; // From session
 
         // Logging values
@@ -223,7 +223,7 @@ const ReportLost = () => {
                         ))}
                     </select>
                     <input className="w-full px-4 py-2 rounded-lg mb-4" type="url" placeholder="Item photo URL" />
-                    <textarea className="w-full px-4 py-2 rounded-lg mb-4" placeholder="Description" required></textarea>
+                    <textarea className="w-full px-4 py-2 rounded-lg mb-4" placeholder="Description" name="description" required></textarea>
                     <button className="w-full px-4 py-2 rounded-lg bg-blue-600 text-white mb-4" type="submit">Submit</button>
                 </form>
                 {submitted && <p className="mt-4 text-green-500">You have successfully reported a lost item.</p>}
